credit approval prediction using ML model
A renowned bank has been facing a challenge: the percentage of credit card customers who defaulted payments has increased from 5% to 9%. This required them to thoroughly review the existing application screening process. Based on the results of the same, they believe that one of the reasons for the increase in defaulters is a lapse in process adherence. This might have been due to a rise in the volume of credit card applications, which is up by 35% in the past three months.

To accurately assess the risk involved and approve credit to financially reliable applicants, the bank wants to introduce a machine learning-based solution. As a part of the Data Science team of the bank, you have been tasked with analyzing the data of past applicants, identifying patterns in it, building an ML model to accurately determine the applicants' credibility, and identifying the attributes that drive eligibility for availing credit cards. The end goal is to use this ML-based solution to streamline the initial screening process to efficiently and accurately assess the creditworthiness of applications.